fix(android): fix touch/keyboard input silently rejected by the bridge
Root cause, confirmed live against a real navetted session: client_id/
surface_id are genuine 64-bit values that can exceed Long.MAX_VALUE (e.g.
15272202610726850855, seen on a real session). MediaInput serialized them
via Kotlin's Long, which JSON-encodes a value with the sign bit set as a
negative decimal literal -- and the bridge's serde deserializer correctly
refuses a '-' sign for a u64 field, rejecting every pointer_motion,
pointer_button and keyboard_key the client sent with "invalid value:
integer `-N`, expected u64". No crash, no visible error: input just did
nothing, video decode was unaffected since it doesn't touch these fields.

Fix: MediaInput's client_id/surface_id are now ULong, which
kotlinx.serialization encodes as the correct unsigned decimal -- exactly
what the bridge expects. StreamConfig's own fields are untouched (Long,
correct for a value only ever echoed within this binary protocol, never
re-serialized to JSON). InputMapper converts at the one boundary that
needs it.

Also replaces the AndroidView's Modifier.pointerInput touch handling with
a direct SurfaceView.setOnTouchListener. This was tried as a fix for what
turned out to be a red herring -- AndroidView's internal
pointerInteropFilter is a real, documented mechanism by which a
pointerInput on/around an AndroidView can silently not fire, but it was
not what was happening here; the actual client_id bug explains the full
symptom under either touch-dispatch approach. Kept anyway since it is now
verified working end to end on a real device and needs no coroutine
gesture-scope ceremony for this slice's single-pointer tracking.

Verified live end-to-end against a real navetted + wprsd + Firefox
session on a Pixel 9 Pro Fold: tapping a button opens a real new browser
session (confirmed via the page navigating to Firefox's own start page),
dragging moves the pointer (confirmed via Firefox's own link-hover status
bar tracking the drag), and the bridge's "invalid_input" rejection is
gone from the logs entirely post-fix.

109 -> 110 unit tests: added a regression test pinning a client_id above
Long.MAX_VALUE serializing as its correct unsigned decimal, using the
exact value observed live. `./gradlew clean assembleDebug
testDebugUnitTest lintDebug` -- the exact CI command -- passes clean.

+++ android/app/src/main/kotlin/com/greponlabs/navette/ui/session/InputMapper.kt
@@ -28,7 +28,7 @@ object InputMapper {
     const val LAYOUT_INDEX: Int = 0
 
     fun pointerMotion(clientId: Long, surfaceId: Long, x: Double, y: Double): MediaInput.PointerMotion =
-        MediaInput.PointerMotion(clientId = clientId, surfaceId = surfaceId, x = x, y = y)
+        MediaInput.PointerMotion(clientId = clientId.toULong(), surfaceId = surfaceId.toULong(), x = x, y = y)
 
     /**
      * Button is always `BTN_LEFT`: this slice maps a tap to a left click and
@@ -36,8 +36,8 @@ object InputMapper {
      */
     fun pointerButton(clientId: Long, surfaceId: Long, pressed: Boolean): MediaInput.PointerButton =
         MediaInput.PointerButton(
-            clientId = clientId,
-            surfaceId = surfaceId,
+            clientId = clientId.toULong(),
+            surfaceId = surfaceId.toULong(),
             button = BTN_LEFT,
             pressed = pressed,
         )
@@ -49,16 +49,16 @@ object InputMapper {
         pressed: Boolean,
     ): MediaInput.KeyboardKey =
         MediaInput.KeyboardKey(
-            clientId = clientId,
-            surfaceId = surfaceId,
+            clientId = clientId.toULong(),
+            surfaceId = surfaceId.toULong(),
             keycode = evdevCode,
             pressed = pressed,
         )
 
     fun keyboardModifiers(clientId: Long, surfaceId: Long, metaState: Int): MediaInput.KeyboardModifiers =
         MediaInput.KeyboardModifiers(
-            clientId = clientId,
-            surfaceId = surfaceId,
+            clientId = clientId.toULong(),
+            surfaceId = surfaceId.toULong(),
             ctrl = metaState and KeyEvent.META_CTRL_ON != 0,
             alt = metaState and KeyEvent.META_ALT_ON != 0,
             shift = metaState and KeyEvent.META_SHIFT_ON != 0,
